Splitted classes in several modules
