+++ src/ntfpy/types/__init__.py
@@ -1,0 +1,4 @@
+from .server  import *
+from .user    import *
+from .client  import *
+from .message import *
