+++ src/ntfpy/types/server.py
@@ -1,0 +1,8 @@
+
+__all__ = [
+    "NTFYServer"
+]
+
+class NTFYServer():
+    def __init__(self, url):
+        self.url = url
